+++ irc3/plugins/autojoins.py
@@ -8,7 +8,7 @@ Auto join channels. The bot will retry to join when kicked and will retry to
 join each 30s when an error occurs.
 
 ..
-    >>> from testing import IrcBot
+    >>> from irc3.testing import IrcBot
 
 Usage::